Move benchmark read models out of control plane
