+++ examples/benchmark-experiment-report-template-smoke.py
@@ -13,7 +13,7 @@ REPO_ROOT = Path(__file__).resolve().parents[1]
 if str(REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(REPO_ROOT))
 
-from loopx.control_plane.runtime.benchmark_comparison import (  # noqa: E402
+from loopx.benchmarks.read_models.benchmark_comparison import (  # noqa: E402
     benchmark_comparison_decision_note,
 )
