+++ loopx/benchmarks/__init__.py
@@ -1,0 +1,1 @@
+"""Benchmark domain code shipped with LoopX."""
