+++ loopx/benchmarks/read_models/__init__.py
@@ -1,0 +1,1 @@
+"""Public-safe benchmark projections and qualification read models."""
