+++ loopx/benchmarks/read_models/benchmark_result.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from .public_safety import (
+from ...control_plane.runtime.public_safety import (
     compact_numeric_map,
     public_safe_compact_list,
     public_safe_compact_text,
