+++ loopx/cli_commands/benchmark_review_lifecycle.py
@@ -22,13 +22,13 @@ from ..benchmark_adapters.terminal_bench import (
     TERMINAL_BENCH_MANAGED_CODEX_LOOPX_KWARGS,
     agent_kwargs_from_invocation,
 )
-from ..control_plane.runtime.benchmark_comparison import (
+from ..benchmarks.read_models.benchmark_comparison import (
     compact_benchmark_comparison,
 )
-from ..control_plane.runtime.benchmark_learning_ledger import (
+from ..benchmarks.read_models.benchmark_learning_ledger import (
     compact_benchmark_learning_ledger,
 )
-from ..control_plane.runtime.benchmark_result import compact_benchmark_result
+from ..benchmarks.read_models.benchmark_result import compact_benchmark_result
 from ..control_plane.work_items.delivery_batch_scale import DELIVERY_BATCH_SCALE_CHOICES
 from ..control_plane.work_items.delivery_outcome import DELIVERY_OUTCOME_CHOICES
 from ..global_registry import sync_project_registry_to_global
